requirements.txt
fastapi==0.89.1
tortoise-orm==0.19.3
pydantic==1.9.0
uvicorn==0.20.0
python-multipart==0.0.6